AutoEncoders/Digits (MLP)/requirements.txt
matplotlib==3.10.1
numpy==1.23.5
pygame==2.6.1
tensorflow-macos==2.10.0
